Providers moved to the common folder
