+++ src/Softville.Upwork.BusinessLogic/Processor/Migrator/OffersMigrator.cs
@@ -2,10 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using Microsoft.Extensions.Logging;
-using Softville.Upwork.BusinessLogic.Processor.ApplicantsStats;
+using Softville.Upwork.BusinessLogic.Processor.Models;
 using Softville.Upwork.BusinessLogic.Processor.Parsers;
+using Softville.Upwork.BusinessLogic.Processor.Providers.ApplicantsStats;
 using Softville.Upwork.BusinessLogic.Processor.Repositories;
-using Softville.Upwork.BusinessLogic.Processor.UpworkApi;
+using Softville.Upwork.BusinessLogic.Processor.Storing;
 
 namespace Softville.Upwork.BusinessLogic.Processor.Migrator;
