+++ src/Softville.Upwork.BusinessLogic/Processor/Providers/ApplicantsStats/ApplicantsStatsMapper.cs
@@ -3,7 +3,7 @@
 
 using Softville.Upwork.Contracts;
 
-namespace Softville.Upwork.BusinessLogic.Processor.ApplicantsStats;
+namespace Softville.Upwork.BusinessLogic.Processor.Providers.ApplicantsStats;
 
 internal static class ApplicantsStatsMapper
 {
