+++ src/Softville.Upwork.BusinessLogic/Processor/Providers/ApplicantsStats/ApplicantsStatsProvider.cs
@@ -3,7 +3,7 @@
 
 using Softville.Upwork.Contracts;
 
-namespace Softville.Upwork.BusinessLogic.Processor.ApplicantsStats;
+namespace Softville.Upwork.BusinessLogic.Processor.Providers.ApplicantsStats;
 
 internal class ApplicantsStatsProvider(IApplicantsClient applicantsClient) : IApplicantsStatsProvider
 {
