+++ src/Softville.Upwork.BusinessLogic/Processor/Providers/OfferDetails/IOfferDetailsUpworkClient.cs
@@ -3,7 +3,7 @@
 
 using Softville.Upwork.Contracts;
 
-namespace Softville.Upwork.BusinessLogic.Processor.OfferDetails;
+namespace Softville.Upwork.BusinessLogic.Processor.Providers.OfferDetails;
 
 public interface IOfferDetailsUpworkClient
 {
