+++ src/Softville.Upwork.BusinessLogic/Processor/Providers/OfferDetails/OfferDetailsProvider.cs
@@ -1,9 +1,10 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using Softville.Upwork.BusinessLogic.Processor.Models;
 using Softville.Upwork.Contracts;
 
-namespace Softville.Upwork.BusinessLogic.Processor.OfferDetails;
+namespace Softville.Upwork.BusinessLogic.Processor.Providers.OfferDetails;
 
 internal class OfferDetailsProvider(IOfferDetailsUpworkClient detailsClient) : IOfferDetailsProvider
 {
