+++ src/Softville.Upwork.BusinessLogic/Processor/Providers/Search/ISearchResultProvider.cs
@@ -1,7 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Softville.Upwork.BusinessLogic.Processor;
+namespace Softville.Upwork.BusinessLogic.Processor.Providers.Search;
 
 internal interface ISearchResultProvider
 {
