+++ src/Softville.Upwork.BusinessLogic/Processor/Storing/IHttpResponseStoring.cs
@@ -1,6 +1,6 @@
 ﻿using Softville.Upwork.Contracts;
 
-namespace Softville.Upwork.BusinessLogic.Processor.UpworkApi;
+namespace Softville.Upwork.BusinessLogic.Processor.Storing;
 
 internal interface IHttpResponseStoring
 {
