+++ src/Softville.Upwork.BusinessLogic/Processor/Storing/IUpworkRequestType.cs
@@ -1,0 +1,9 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Softville.Upwork.BusinessLogic.Processor.Storing;
+
+public interface IUpworkRequestType
+{
+    public string RequestName { get; }
+}
