+++ tests/Softville.Upwork.BusinessLogic.Tests/Processor/Mappers/UpworkApplicantsStatsMapper.cs
@@ -1,8 +1,8 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Softville.Upwork.BusinessLogic.Processor.ApplicantsStats;
 using Softville.Upwork.BusinessLogic.Processor.Parsers;
+using Softville.Upwork.BusinessLogic.Processor.Providers.ApplicantsStats;
 using Softville.Upwork.Tests.Common.Data;
 
 namespace Softville.Upwork.BusinessLogic.Tests.Processor.Mappers;
