+++ tests/Softville.Upwork.BusinessLogic.Tests/Processor/Mappers/UpworkOfferMapper.cs
@@ -1,7 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Softville.Upwork.BusinessLogic.Processor;
+using Softville.Upwork.BusinessLogic.Processor.Models;
 using Softville.Upwork.BusinessLogic.Processor.Parsers;
 using Softville.Upwork.Tests.Common.Data;
