+++ tests/Softville.Upwork.Tests.Common/Softville.Upwork.Tests.Common.csproj
@@ -18,6 +18,7 @@
       <EmbeddedResource Include="Data\upwork-applicantsmodel.json" />
       <EmbeddedResource Include="Data\upwork-fulldatamodel-v2.json" />
       <EmbeddedResource Include="Data\upwork-fulldatamodel.json" />
+      <EmbeddedResource Include="Data\upwork-searchresultmodel-v2.json" />
       <EmbeddedResource Include="Data\upwork-searchresultmodel.json" />
     </ItemGroup>
